Collision reporting
We can now listen to collisions that occur for an entity.

For example if we want to know when the player collies with a specific
entity by ID we can listen for this event to occur.

To listen to a collision event
Send message -  ListenToCollision(int entityID)

This message is then created when a collision event occurs
CollisionReport();
- The message stores data of the two entities that collided.